support json configuration file and refactor the code.
