+++ cmake/presets/ort_genai.cmake
@@ -5,3 +5,5 @@ set(OCOS_ENABLE_GPT2_TOKENIZER ON CACHE INTERNAL "" FORCE)
 set(OCOS_ENABLE_C_API ON CACHE INTERNAL "" FORCE)
 set(OCOS_ENABLE_CV2 ON CACHE INTERNAL "" FORCE)
 set(OCOS_ENABLE_OPENCV_CODECS ON CACHE INTERNAL "" FORCE)
+
+set(OCOS_ENABLE_CTEST OFF CACHE INTERNAL "" FORCE)
