+++ shared/api/image_transforms.hpp
